requirements.txt
fastapi
uvicorn 
sqlmodel 
sqlalchemy 
pydantic
pydantic-settings
pyjwt 
passlib
pyodbc
python-jose
python-dotenv